Extract a Monitor object to get stats of the Audit
We shouldn’t be calculating things on the views/helpers. Views should be
pretty simple and just do some formatting.

In this commit I am extracting the monitor object into the domain layer.
